Move SI_GCAdapter over to InputCommon
