+++ Source/Core/Core/Core.vcxproj.filters
@@ -720,9 +720,6 @@
     <ClCompile Include="PowerPC\MMU.cpp">
       <Filter>PowerPC</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_GCAdapter.cpp">
-      <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
-    </ClCompile>
     <ClCompile Include="PowerPC\JitCommon\JitBackpatch.cpp">
       <Filter>PowerPC\JitCommon</Filter>
     </ClCompile>
@@ -1240,9 +1237,6 @@
     </ClInclude>
     <ClInclude Include="HW\GCKeyboard.h">
       <Filter>HW %28Flipper/Hollywood%29\GCKeyboard</Filter>
-    </ClInclude>
-    <ClInclude Include="HW\SI_GCAdapter.h">
-      <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
     <ClInclude Include="PowerPC\Jit64Common\Jit64AsmCommon.h">
       <Filter>PowerPC\Jit64Common</Filter>
