+++ Source/Core/InputCommon/GCAdapter.cpp
