+++ Source/Core/InputCommon/GCAdapter.h
